move fork and create projectile builders to their own file
