+++ core/core-api/src/main/java/org/kontinuity/catapult/core/api/ForkProjectile.java
@@ -22,7 +22,7 @@ public class ForkProjectile extends Projectile {
     *
     * @param builder
     */
-   ForkProjectile(ProjectileBuilder.ForkProjectileBuilder builder) {
+   ForkProjectile(ForkProjectileBuilder builder) {
       super(builder);
       this.sourceGitHubRepo = builder.getSourceGitHubRepo();
       this.pipelineTemplatePath = builder.getPipelineTemplatePath();
