+++ core/core-api/src/main/java/org/kontinuity/catapult/core/api/ProjectileBuilder.java
@@ -1,9 +1,5 @@
 package org.kontinuity.catapult.core.api;
 
-import java.io.File;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 /**
  * DSL builder for creating {@link Projectile} objects.  Responsible for
  * validating state before calling upon the {@link ForkProjectileBuilder#build()}
@@ -27,14 +23,11 @@ public class ProjectileBuilder {
     */
    private String openShiftProjectName;
 
-
-   private static final Pattern REPO_PATTERN = Pattern.compile("^[a-zA-Z_0-9\\-]+/[a-zA-Z_0-9\\-]+");
-
    private ProjectileBuilder() {
       // No external instances
    }
 
-   private ProjectileBuilder(String gitHubAccessToken, String openShiftProjectName) {
+   ProjectileBuilder(String gitHubAccessToken, String openShiftProjectName) {
       this.gitHubAccessToken = gitHubAccessToken;
       this.openShiftProjectName = openShiftProjectName;
    }
@@ -56,7 +49,7 @@ public class ProjectileBuilder {
     * @param value
     * @throws IllegalStateException
     */
-   private static void checkSpecified(final String name,
+   static void checkSpecified(final String name,
                                       final String value) throws IllegalStateException {
       assert name != null && !name.isEmpty() : "name is required";
       if (value == null || value.isEmpty()) {
@@ -68,7 +61,7 @@ public class ProjectileBuilder {
      * Builder methods
      */
 
-   private void build(ProjectileBuilder builder) {
+   void build(ProjectileBuilder builder) {
       ProjectileBuilder.checkSpecified("gitHubAccessToken", this.gitHubAccessToken);
       // Default the openshiftProjectName if need be
       try {
@@ -129,153 +122,4 @@ public class ProjectileBuilder {
       return new ForkProjectileBuilder(getGitHubAccessToken(), getOpenShiftProjectName());
    }
 
-   public class CreateProjectileBuilder extends ProjectileBuilder {
-      private String projectLocation;
-
-      CreateProjectileBuilder(String gitHubAccessToken, String openShiftProjectName) {
-         super(gitHubAccessToken, openShiftProjectName);
-      }
-
-      /**
-       * Creates and returns a new {@link CreateProjectile} instance based on the
-       * state of this builder; if any preconditions like missing properties
-       * or improper values exist, an {@link IllegalStateException} will be thrown
-       *
-       * @return the created {@link Projectile}
-       * @throws IllegalStateException
-       */
-      public Projectile build() {
-         super.build(this);
-         ProjectileBuilder.checkSpecified("projectLocation", this.projectLocation);
-         return new CreateProjectile(this);
-      }
-
-      @Override
-      String createDefaultProjectName() {
-         return projectLocation.substring(projectLocation.lastIndexOf(File.separator));
-      }
-
-      /**
-       * Sets the projectLocation of the repository this
-       * is what will be "uploaded" for the user.  Required.
-       *
-       * @param projectLocation
-       * @return This builder
-       */
-      public CreateProjectileBuilder projectLocation(final String projectLocation) {
-         this.projectLocation = projectLocation;
-         return this;
-      }
-
-      /**
-       * @return the location of the project to "upload" to GitHub.
-       */
-      public String getProjectLocation() {
-         return projectLocation;
-      }
-   }
-
-   public class ForkProjectileBuilder extends ProjectileBuilder {
-      private String sourceGitHubRepo;
-
-      /**
-       * the path to the file in the repo that contains the pipeline template.
-       */
-      private String pipelineTemplatePath;
-
-      private String gitRef;
-
-      ForkProjectileBuilder(String gitHubAccessToken, String openShiftProjectName) {
-         super(gitHubAccessToken, openShiftProjectName);
-      }
-
-      /**
-       * Creates and returns a new {@link ForkProjectile} instance based on the
-       * state of this builder; if any preconditions like missing properties
-       * or improper values exist, an {@link IllegalStateException} will be thrown
-       *
-       * @return the created {@link Projectile}
-       * @throws IllegalStateException
-       */
-      public Projectile build() throws IllegalStateException {
-         super.build(this);
-         // Precondition checks
-         ProjectileBuilder.checkSpecified("sourceGitHubRepo", this.sourceGitHubRepo);
-         final Matcher matcher = REPO_PATTERN.matcher(sourceGitHubRepo);
-         if (!matcher.matches()) {
-            throw new IllegalStateException("source repo must be in form \"owner/repoName\"");
-         }
-         ProjectileBuilder.checkSpecified("pipelineTemplatePath", this.pipelineTemplatePath);
-         ProjectileBuilder.checkSpecified("girRef", this.gitRef);
-
-         // All good, so make a new instance
-         return new ForkProjectile(this);
-      }
-
-      @Override
-      String createDefaultProjectName() {
-         final String sourceGitHubRepo = this.getSourceGitHubRepo();
-         final String targetProjectName = this.getSourceGitHubRepo().substring(
-               sourceGitHubRepo.lastIndexOf('/') + 1);
-
-         return targetProjectName;
-      }
-
-      /**
-       * Sets the source GitHub repository name in form "owner/repoName"; this
-       * is what will be forked on behalf of the user.  Required.
-       *
-       * @param sourceGitHubRepo
-       * @return This builder
-       */
-      public ForkProjectileBuilder sourceGitHubRepo(final String sourceGitHubRepo) {
-         this.sourceGitHubRepo = sourceGitHubRepo;
-         return this;
-      }
-
-      /**
-       * Sets the path to file that contains the template to apply on the
-       * OpenShift project. Required.
-       *
-       * @param pipelineTemplatePath
-       * @return This builder
-       */
-      public ForkProjectileBuilder pipelineTemplatePath(final String pipelineTemplatePath) {
-         this.pipelineTemplatePath = pipelineTemplatePath;
-         return this;
-      }
-
-
-      /**
-       * Sets Git ref to use. Required
-       *
-       * @param gitRef
-       * @return This builder
-       */
-      public ForkProjectileBuilder gitRef(final String gitRef) {
-         this.gitRef = gitRef;
-         return this;
-      }
-
-      /**
-       * @return source GitHub repository name in form "owner/repoName".
-       */
-      public String getSourceGitHubRepo() {
-         return this.sourceGitHubRepo;
-      }
-
-      /**
-       * @return the path to the file that contains the template to apply on the OpenShift project.
-       */
-      public String getPipelineTemplatePath() {
-         return this.pipelineTemplatePath;
-      }
-
-      /**
-       * @return The Git reference to use
-       */
-      public String getGitRef() {
-         return gitRef;
-      }
-   }
 }
